Resolved all needed changes
